fix typedef in cython code

+++ xtpwrapper/cppheader/c_quote_wrapper.h
@@ -1,14 +1,9 @@
-﻿
-
-#ifndef C_QUOTE_WRAPPER
+﻿#ifndef C_QUOTE_WRAPPER
 #define C_QUOTE_WRAPPER
 
 #include "Python.h"
 #include "pythread.h"
 #include "xtp_quote_api.h"
-
-
-
 
 using namespace XTP::API;
 
@@ -18,7 +13,7 @@ static inline int QuoteSpi_OnError(PyObject *, XTPRI *);
 
 static inline int QuoteSpi_OnSubMarketData(PyObject *, XTPST *, XTPRI *, bool);
 static inline int QuoteSpi_OnUnSubMarketData(PyObject *, XTPST *, XTPRI *, bool);
-static inline int QuoteSpi_OnDepthMarketData(PyObject*, XTPMD*, long long[], int, int, long long[], int, int);
+static inline int QuoteSpi_OnDepthMarketData(PyObject *, XTPMD *, int64_t[], int32_t, int32_t, int64_t[], int32_t, int32_t);
 
 static inline int QuoteSpi_OnSubOrderBook(PyObject *, XTPST *, XTPRI *, bool);
 static inline int QuoteSpi_OnUnSubOrderBook(PyObject *, XTPST *, XTPRI *, bool);
